v5.0.0 (#96)
Major updates
* Using latest Alpine, Busybox and Nushell base images

Minor updates
* Updating to Debian 13.2
* Making record formatting consistent
* Using new clean override

Build updates
* Using quay.io
* Updating action versions on workflows

+++ overlay/etc/nu/scripts/bf/clean.nu
@@ -3,17 +3,27 @@ use fs.nu
 use write.nu
 
 # clean temporary directories, caches and installation files
-export def main []: nothing -> nothing {
-    write debug "Deleting caches." clean
-    rm --force --recursive /tmp/* /var/lib/apt/lists/* /usr/share/man
+export def main [
+    --caches: list<string> = [ "/var/lib/apt/lists/*" "/usr/share/man" ]  # list of cache directories to clean
+    --tmpdirs: list<string> = [ "/tmp/*" ]                                # list of temporary directories to clean
+]: nothing -> nothing {
+    # create a merged list of places to clean -
+    # appending /* to each so we don't delete the actual directories
+    let ensure_glob = $caches
+        | append $tmpdirs
+        | into glob
+    write debug $"Deleting ($ensure_glob | str join ', ')." clean
+    rm --force --recursive ...$ensure_glob
 
+    # search for .empty files
     let files = fs find_name "/" ".empty"
     if ($files | is-not-empty) {
         write debug "Deleting .empty files." clean
         rm --force ...$files
     }
 
-    if (env check PUBLISHING) {
+    # remove installation files from published images
+    if (env check PUBLISHING) and not (env check TESTING) {
         write debug "Deleting preinstallation script." clean
         rm --force /preinstall
 
@@ -21,6 +31,6 @@ export def main []: nothing -> nothing {
         rm --force --recursive /etc/nu/scripts/tests
     }
 
-    write ok "Done." clean
+    # return nothing
     return
 }
